db: Use fedora postgres image
Using the official fedora postgres image. Its very easy to configure
and start with. On the initial run it will try to create a db
under its data directory.

Running this container with volume mounted on /var/lib/pgsql/data will
invoke initdb on it, creating a nested data/ dir under it with postgres
password defaulting to 'postgres'.

After running the compose we can create the engine db from the developer
container:
```
  psql -h db postgres postgres <<EOF
    create user engine password 'engine';
    create database engine owner engine template template0
    encoding 'UTF8' lc_collate 'en_US.UTF-8' lc_ctype 'en_US.UTF-8';
  EOF
```

Signed-off-by: Roy Golan <rgolan@redhat.com>